src/main/webapp/app/entities/reward/route/reward-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IReward, Reward } from '../reward.model';
import { RewardService } from '../service/reward.service';

import { RewardRoutingResolveService } from './reward-routing-resolve.service';

describe('Reward routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: RewardRoutingResolveService;
  let service: RewardService;
  let resultReward: IReward | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(RewardRoutingResolveService);
    service = TestBed.inject(RewardService);
    resultReward = undefined;
  });

  describe('resolve', () => {
    it('should return IReward returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultReward = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultReward).toEqual({ id: 123 });
    });

    it('should return new IReward if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultReward = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultReward).toEqual(new Reward());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Reward })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultReward = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultReward).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
